refactor(mcp): extract neutral WeveClient core; address PR #251 review
Split the WeveNova MCP client so the planner no longer subclasses the
landing-page AgentConfigClient. A neutral `weve/` core (WeveClient +
shared OData helpers) is now composed as siblings by `agentconfig/`
(AgentConfigClient) and `planner/` (PlannerClient) under src/mcp/.

- weve/weve_client.py: neutral core (auth, token decode, shared httpx
  session, retrying _request) with a _transform_response identity hook.
- weve/_odata.py: shared helpers; dedup _validate_odata_string,
  _require_odata_id and _validate_https_base_url; JWT decode moved to core.
- agentconfig/client.py: slim AgentConfigClient(WeveClient) keeping the
  EmployeeAgents routes and PascalCase payload transform.
- planner/: PlannerClient(PlannerMixin, RolesMixin, WeveClient); server
  renamed server_planner.py -> server.py.
- Retry: default idempotent=True; non-convergent planner creates
  (plan / task / role-task) opt out, while get-or-create project and
  attest stay retry-safe.
- list_project_plan_tasks_for_caller expands to the caller's active role
  assignments so role-pooled tasks are no longer dropped.
- roles.py: fix the stale class-attr comment to reference PlannerClient.
- Tests relocated to tests/mcp/{weve,planner}; add a landing-page create
  retry test and caller role-expansion tests.
- Register ess-planner in mcp.defaults.json; both servers share one
  weve/.local MSAL token cache and interactive sign-in.
