+++ solutions/ess-maker-skills/src/mcp/planner/requirements.txt
@@ -1,0 +1,3 @@
+mcp>=1.29.0,<2.0.0
+httpx>=0.27.0,<1.0
+msal>=1.35.0
