+++ solutions/ess-maker-skills/src/mcp/weve/requirements.txt
@@ -1,0 +1,2 @@
+httpx>=0.27.0,<1.0
+msal>=1.35.0
